requirements.txt
fuzzywuzzy==0.18.0
PyYAML==6.0
requests==2.26.0
